fix: delete 시 userId 를 검색에 포함하도록 수정

+++ src/test/java/until/the/eternity/dcs/infrastructure/FakeCommentLikeRepository.java
@@ -39,9 +39,9 @@ public class FakeCommentLikeRepository implements CommentLikeRepository {
 	}
 
 	@Override
-	public void deleteByCommentId(Long commentId) {
+	public void deleteByCommentIdAndUserId(Long commentId, Long userId) {
 		db.values().forEach(commentLike -> {
-			if (commentLike.getCommentId().equals(commentId)) {
+			if (commentLike.getCommentId().equals(commentId) && commentLike.getUserId().equals(userId)) {
 				db.remove(commentLike.getId());
 			}
 		});
